% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sherbert_colors.R
\name{sherbert_pal}
\alias{sherbert_pal}
\title{Return function to interpolate an sherbert color palette}
\usage{
sherbert_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in sherbert_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an sherbert color palette
}
